+++ bg_project/rl_games/__init__.py
@@ -1,0 +1,6 @@
+from gymnasium.envs.registration import register
+
+register(
+    id="GridWorld-v0",
+    entry_point="rl_games.envs:GridWorld",
+)
