+++ bg_project/rl_games/envs/grid_worlds.py
@@ -1,0 +1,15 @@
+from gymnasium import Env
+
+
+class GridWorld(Env):
+    def __init__(self):
+        super().__init__()
+
+    def reset(self):
+        pass
+
+    def step(self):
+        pass
+
+    def render(self):
+        pass
